requirements.txt
# =========================
# FastAPI core
# =========================
fastapi==0.128.5
starlette==0.52.1
uvicorn[standard]==0.40.0

# =========================
# Database (Postgres / Yugabyte)
# =========================
SQLAlchemy==2.0.46
alembic==1.18.3
asyncpg==0.30.0
greenlet==3.3.1

# =========================
# Config / env
# =========================
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

# =========================
# Auth / Security
# =========================
PyJWT==2.11.0
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==46.0.4

# =========================
# HTTP client (Ollama + Qdrant)
# =========================
httpx==0.28.1
httpcore==1.0.9

# =========================
# File upload support
# =========================
python-multipart==0.0.9

# =========================
# PDF extraction (RAG ingestion)
# =========================
pypdf==4.3.1

# =========================
# Validation
# =========================
email-validator==2.3.0

# =========================
# Templates
# =========================
Jinja2==3.1.6
MarkupSafe==3.0.3

# =========================
# Seed / utilities
# =========================
Faker==40.4.0
tqdm==4.67.3

# =========================
# YAML support
# =========================
PyYAML==6.0.3

# =========================
# typing support
# =========================
typing_extensions==4.15.0

# =========================
# uvicorn runtime deps
# =========================
anyio==4.12.1
h11==0.16.0
httptools==0.7.1
watchfiles==1.1.1
websockets==16.0

# =========================
# timezone support
# =========================
tzdata==2025.3
